Implemented ω-CPOs and a lifting operator.
Using a higher inductive-inductive construction, following a suggestion
from Paolo Capriotti that came up in a discussion that also involved
Thorsten Altenkirch and Bernhard Reus (but not myself).

I also defined the partiality monad using the lifting operator, and
proved that this definition is equivalent to the one in
Partiality-monad.Inductive.