Update in response to changes to Agda's option handling.
The sized types primitives can not be declared unless --sized-types is
active, and when the --safe option is used sized types are now by
default turned off. I don't want to activate --sized-types in the
Prelude module (because then it cannot be used by code which uses
--safe and --guardedness), so the primitives are now reexported from a
new module, Size.